:recycle: Refactor directories
