+++ lib/domain/calendar/weekly/calendar_tab_state.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pilll/domain/calendar/date_range.dart';
-import 'package:pilll/domain/calendar/weekly/weekly_calendar_state.dart';
+import 'package:pilll/components/organisms/calendar/weekly/weekly_calendar_state.dart';
 import 'package:pilll/entity/diary.dart';
 
 class CalendarTabWeeklyCalendarState extends WeeklyCalendarState {
